Added plugin framework
Added plugin framework which allows to create communication backends as
loadable modules.

For example, the included USB backend is now installed in eg.:
/usr/lib/sensact/plugins/sensact-usb.so

To load it simply do:
sa_plugin_load("usb");

The filename of the plugin must be of the sensact-<name>.so format.

Once loaded, you can register your devices which uses the USB backend.

Actually, with the new plugin infrastructure one can also choose to
register devices from a plugin. So, it is possible to have backend
and/or device configuration within the plugin (libsensact does not
implement policy so you decide).